cheri: Increase context size of skb, netlink and textsearch
These contexts are used to store caller specific data and on
CHERI the size of this data may be larger if it contains pointers.

For the netlink callback data the args array holds capabilities, now.
Adjust users as necessary.

+++ include/net/cfg80211.h
@@ -5642,7 +5642,7 @@ struct wiphy_vendor_command {
 		    const void *data, int data_len);
 	int (*dumpit)(struct wiphy *wiphy, struct wireless_dev *wdev,
 		      struct sk_buff *skb, const void *data, int data_len,
-		      unsigned long *storage);
+		      uintptr_t *storage);
 	const struct nla_policy *policy;
 	unsigned int maxattr;
 };
